update/ rename SearchResult component
