+++ pages/search/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SearchResults from "@/modules/Search";
+import SearchResults from "@/modules/SearchResults";
 
 export default function SearchPage() {
   return (
